requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy[asyncio]
aiosqlite
asyncpg
strawberry-graphql[fastapi]
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
pydantic-settings
python-multipart
alembic
httpx
redis
beautifulsoup4
lxml
Pillow
slowapi
playwright
anthropic
stripe
python-dateutil
nh3
cryptography
